core: Put functions acting on the local particle index into separate file
