+++ src/core/CMakeLists.txt
@@ -24,6 +24,7 @@ set(EspressoCore_SRC
     nsquare.cpp
     partCfg_global.cpp
     particle_data.cpp
+    particle_index.cpp
     polymer.cpp
     polynom.cpp
     pressure.cpp
