+++ src/core/domain_decomposition.cpp
@@ -28,7 +28,7 @@
 #include "communication.hpp"
 #include "errorhandling.hpp"
 #include "grid.hpp"
-#include "particle_data.hpp"
+#include "particle_index.hpp"
 
 #include "serialization/ParticleList.hpp"
 #include <utils/index.hpp>
