+++ src/core/forces_inline.hpp
@@ -60,6 +60,7 @@
 #include "object-in-fluid/oif_global_forces.hpp"
 #include "object-in-fluid/oif_local_forces.hpp"
 #include "particle_data.hpp"
+#include "particle_index.hpp"
 #include "rotation.hpp"
 #include "thermostat.hpp"
