+++ src/core/object-in-fluid/oif_global_forces.cpp
@@ -25,7 +25,7 @@
 #include "errorhandling.hpp"
 #include "grid.hpp"
 #include "grid_based_algorithms/lb_interface.hpp"
-#include "particle_data.hpp"
+#include "particle_index.hpp"
 
 #include <utils/math/triangle_functions.hpp>
 using Utils::angle_btw_triangles;
